The function ChooseAttribute
Finished the most of the function ChooseAttribute but have not tested it yet. I will try to run a demonstration and have a look for other parts on Thursday. If you get any problems with the details of the function or the value "activeAttribute" please send messages to me, or we can discuss it on Friday.